ForestFloorExport {Rborist}	R Documentation

Exportation Format for Rborist Training Output

Description

Formats training output into a form suitable for forestFloor feature-contribution package.

Usage

 ## S3 method for class 'Rborist'
ForestFloorExport(arbOut)

Arguments

arbOut
an object of type Rborist produced by training.

Value

An object of type ForestFloorExport, as specified by the interface for the forestFloor package.

Author(s)

Mark Seligman at Suiji.

Examples

  ## Not run: 
    data(iris)
    rb <- Rborist(iris[-5], iris[5])
    ffe <- ForestFloorExport(rb)

    library(forestFloor)
    ForestFloor(ffe)
  
## End(Not run)

[Package Rborist version 0.1-3 Index]